Move the response content type to a more general namespace
Follow-up to #15
